+++ packages/sign-client/test/shared/index.ts
@@ -4,3 +4,4 @@ export * from "./init";
 export * from "./mock";
 export * from "./values";
 export * from "./helpers";
+export * from "./metrics";
